Added ability to copy games to OPL folder from a menu option.
This is currently an incomplete implementation, it needs to be threaded and display a progress bar and it also needs to have the file select filter applied for only ISO files.

It also relies on the cluster size to detect the file type which isn't entirely accurate but there's little information on how to detect CD vs DVD ISOs. This is the best we've got.

For now this is an initial implementation.